requirements.txt
discord.py
python-dotenv
google-generativeai>=0.8.0